Style profile page and show srt
Finish styling the structure of profile page, but the function of
request student's all lessons from backend haven't done yet.

Use searchSrt from "api/search" to get srt of keyword from backend.
Using map() function to show all the lessons that include the
keyword, also show the srt include the keyword by <Srt /> in
<VideoCard />.

Found a bug that if you refresh the page which is not HomePage,
the Navbar will show Login button instead of user icon while in
the login status. Fixing this bug by moving useEffect(), which
is in HomePage/index.js, to component/Navbar, and clear some
unnecessary imports and declarations in HomePage/index.js and
App.js.