Generate response to user's question in Chatbot
Create api chatbot.js and use this api to fetch response that
generated by model.

Modify react-simple-chatbot to feat the style of our project
and use useEffect() to put the response into botBubble.

animation.jsx and loading.jsx is a loader for waiting response
generattion. When the response finish generate and the state was
set, the loader will be change to the generated response.