Direct to VideoPage while clicking srts
Use <div> to replace <a> because there will cause some warning
while using <a>,so I create a function handleSrtClick() to
navigate to the VideoPage with lesson_id when clicking srt in
srt block, haven't sent timestamp to VideoPage yet.

Create a loader animation while waiting the generation of search
summary.

Fix the problem of changing 'class' to 'className' in LoginModal
and Navbar.

Add navigate(0) when directing to SearchPage after typing keyword
and click send. Because if I enter another keyword in SearchPage,
the content of search summary won't change, I think this will
cause bad user experience, so I add navigate(0) to refresh the
page after sending the keyword.

Add navigate(0) when direct to VideoPage. Right after login, if I
click the img direct to VideoPage, it doesn't fetch the jwt
authentication, but after clicking refresh, it fetched, and I
don't know why. So I add a navigate(0) to refresh the page after
direct to VideoPage in order to fetch jwt authentication.