Set Video watch time to user's timestamp
While navigate to VideoPage, it will send two states, lesson_id
and course_id, and VideoPage will get these two states by using
useLocation.

Find a built-in instance function seekTo() in react-player, it
can set the video to the time or portion you want with useRef, so
I comment the state videoTime because it doesn't necessary.