Revert "Set Video watch time to user's timestamp"
This reverts commit e405406a1adf450aabd3b40c5da096bf228a916a.

I try to write a new api and a new function to fetch the summary
of the lesson that was in search/srt, but eventually I found that
it will cause error of affecting showing srt normally, and I'm
lazy to fix it. So, I push the files except of SearchPage to
GitHub, remove frontend and clone the brand new frontend to my
laptop. Right after I make some change and try to run, it show
up some errors, and then I start fixing it. However, after hours
and hours, I still can't figure out where cause the error. So, I
decide to revert it because of my dumb manipulation.