Stick the Navbar at the top and remove sidebar
Use position: fixed and set extra space to let the navbar to stick
at the top of the webpage. Use position fixed instead of sticky
is because position: sticky isn't work as well, and I can't find
solutions to fix it, so now just use position fixed.

Because Sidebar is redundant, so I remove it.